Add files via upload
This code is a reverse-engineered templated stl doubly-linked list class written in C++. It was written for an assignment to mimic the C++ list standard template library class. This can be used in place of the C++ stl list class. 

One significant difference is that my iterator class will be a templated class of its own, not a class within the list class.  So to get a list of cars, you would say:  list<cars> mylist;  However, there is a built-in class called iterator (different from the list<..>::iterator class).  So, we cannot use iterator as the name of our iterator class.  It uses iter.  To get an iterator to a node within a list of cars, you would use: iter<cars> myiterator;